v0.69
Время бежит, а мы должны еще быстрее. В этом выпуске прошивки добавлены
функции работы со временем.

- Доработан класс драйвера времени. Добавлены функции синхронизации с
сервером. При старте устройства происходит запрос на получение времени с
сервера и запись его в рамках класса.
- HTTPClient: В случае, если IP не задан, при запросе прошивка пытается
его получить. Если IP не может быть найден - вызывается task_fatal_error
и выполнение процесса останавливается.
- В раздел API /device добавлено отображение текущего времени устройства
и часового пояса
- В раздел Api /device добавлено сохранение часового пояса пользователя.
 Часовой пояс сохраняется классом drivers/time в памяти устройства и
восстанавливается после перезапуска.
- Добавлена возможность задания текущего времени через POST запрос для
сохранения этих настроек через мобильное приложение

+++ src/drivers/NVS/NVS.cpp
@@ -162,7 +162,7 @@ void NVS::BlobArrayRemove(string ArrayName, uint8_t Index) {
 	Erase(ArrayName + "_" + Tools::ToString(ArrayCount));
 
 	for (uint8_t i = Index; i < (ArrayCount-1) ; i++) {
-		void * tmp = GetBlob(ArrayName + "_" + Tools::ToString(i+1));
+		void * tmp = GetBlob(ArrayName + "_" + Tools::ToString((uint8_t)(i+1)));
 		SetBlob(ArrayName + "_" + Tools::ToString(i), tmp);
 	}
 
@@ -218,11 +218,11 @@ void NVS::StringArrayRemove(string ArrayName, uint8_t Index) {
 	Erase(ArrayName + "_" + Tools::ToString(ArrayCount));
 
 	for (uint8_t i = Index; i < (ArrayCount-1) ; i++) {
-		string tmp = GetString(ArrayName + "_" + Tools::ToString(i+1));
-		SetString(ArrayName + "_" + Tools::ToString(i), tmp);
+		string tmp = GetString(ArrayName + "_" + Tools::ToString((uint8_t)(i+1)));
+		SetString(ArrayName + "_" + Tools::ToString((uint8_t)i), tmp);
 	}
 
-	Erase(ArrayName + "_" + Tools::ToString(ArrayCount));
+	Erase(ArrayName + "_" + Tools::ToString((uint8_t)ArrayCount));
 	ArrayCountSet(ArrayName, (ArrayCount-1));
 }
